fix: CI workflow (#95)
<!--
Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
    have the right to submit it under the open source license
    indicated in the file; or

(b) The contribution is based upon previous work that, to the best
    of my knowledge, is covered under an appropriate open source
    license and I have the right under that license to submit that
    work with modifications, whether created in whole or in part
    by me, under the same open source license (unless I am
    permitted to submit under a different license), as indicated
    in the file; or

(c) The contribution was provided directly to me by some other
    person who certified (a), (b) or (c) and I have not modified
    it.

(d) I understand and agree that this project and the contribution
    are public and that a record of the contribution (including all
    personal information I submit with it, including my sign-off) is
    maintained indefinitely and may be redistributed consistent with
    this project or the open source license(s) involved.
-->

I have fixed the CI workflow errors.
And I changed the publish method to [Trusted
Publishing](https://crates.io/docs/trusted-publishing). You just need to
do some simple settings to make it work:
1. Open your crate: https://crates.io/crates/cargo-tag
2. Click `Settings` tab
3. Click `Add` button next to `Trusted Publishing`
4. Fill `Repository owner` with LeoBorai
5. Fill `Repository name` with cargo-tag
6. Fill `Workflow filename` with release.yml
7. Click `Add`