fix(git): support branch discovery (#112)
This pull request makes a small change to how the repository is opened
in the `Git` implementation. Instead of using the current working
directory, it now uses `Repository::open_from_env()`, which is more
robust and respects environment variables for repository discovery.

* Replaced usage of `current_dir()` and `Repository::open(cwd)` with
`Repository::open_from_env()` in the `Git::open` method for improved
repository discovery.
* Removed the unnecessary import of `current_dir` from `std::env`.